Update .eslintrc.js
/src/renderer/plugins/vuetify.js imports fonts for offline use. 

Eslint detects these as "unused-expressions".

This causes the initial `yarn` command to fail and exit with code 1.

Although it failed, a `yarn run dev` still works.

Adding this line, the command initial `yarn` will work as expected.
Maybe there is a better solution here. Globally disabling this rule can have other side effects. They won't be as hard as we're not in a size-critical environment.

Hope this helps.